begin refactoring of pages and widgets to display API infos
