+++ lib/ui/pages/dashboard.dart
@@ -1,29 +1,34 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:cesi_covid_19_tracker/data/models/covid_country_infos.dart';
 import 'package:flutter/material.dart';
 
-import 'package:cesi_covid_19_tracker/ui/widgets/widgets.dart'
-    show FailureIcon, NavigationDrawer;
 import 'package:cesi_covid_19_tracker/data/services/locator.dart';
 import 'package:cesi_covid_19_tracker/data/services/services.dart'
     show AppUtils;
+import 'package:cesi_covid_19_tracker/data/models/covid_infos.dart';
+import 'package:cesi_covid_19_tracker/ui/widgets/widgets.dart'
+    show CoronedGlobalCard, FailureIcon, NavigationDrawer;
 import 'package:cesi_covid_19_tracker/data/constants/app_globals.dart' as aG;
 
-class DashBoard extends StatefulWidget {
+class Dashboard extends StatefulWidget {
+  final String title;
+
+  Dashboard({
+    Key key,
+    this.title = aG.AppConstants.defaultAppTitle,
+  }) : super(key: key);
   @override
-  _DashBoardState createState() => _DashBoardState();
+  _DashboardState createState() => _DashboardState();
 }
 
-class _DashBoardState extends State<DashBoard> {
-  String _dropDownValue;
-  String _apiResponse;
-  StreamController<String> _apiResponseController;
+class _DashboardState extends State<Dashboard> {
+  StreamController _apiResponseController;
 
   @override
   void initState() {
     super.initState();
     _apiResponseController = StreamController();
+    callApi();
   }
 
   @override
@@ -33,11 +38,11 @@ class _DashBoardState extends State<DashBoard> {
       appBar: AppBar(
         title: RichText(
           text: TextSpan(
-            text: aG.AppConstants.defaultAppTitle.split('\t\t')[0],
+            text: widget.title.split('\t\t')[0],
             style: Theme.of(context).textTheme.headline1,
             children: [
               TextSpan(
-                text: '\t\t' + aG.AppConstants.defaultAppTitle.split('\t\t')[1],
+                text: '\t\t' + widget.title.split('\t\t')[1],
                 style: Theme.of(context).textTheme.bodyText2,
               ),
             ],
@@ -51,39 +56,108 @@ class _DashBoardState extends State<DashBoard> {
         primary: true,
         shrinkWrap: true,
         children: <Widget>[
-          SingleChildScrollView(
-            child: Column(
-              children: _buildChildren()
-                ..add(
-                  StreamBuilder<String>(
-                    stream: _apiResponseController.stream,
-                    builder: (_, AsyncSnapshot<String> s) {
-                      print('Has error: ${s.hasError}');
-                      print('Has data: ${s.hasData}');
-                      print('Snapshot Data ${s.data}');
-                      if (s.hasError) {
-                        return FailureIcon(fail: s.error);
-                      }
-                      if (s.hasData) {
-                        var cCL =
-                            CovidCountryInfos.fromJson(jsonDecode(s.data));
-                        return Text(
-                          cCL.toString(),
-                          style: TextStyle(color: Colors.black),
-                        );
-                      }
-                      if (s.connectionState != ConnectionState.done) {
-                        return Container();
-                      }
-                      if (!s.hasData &&
-                          s.connectionState == ConnectionState.done) {
-                        return FailureIcon(fail: 'No Data');
-                      } else {
-                        return Container();
-                      }
-                    },
-                  ),
-                ),
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 24.0),
+            child: StreamBuilder(
+              stream: _apiResponseController.stream,
+              builder: (_, s) {
+                print('Has error: ${s.hasError}');
+                print('Has data: ${s.hasData}');
+                print('Snapshot Data ${s.data}');
+
+                if (s.hasData) {
+                  var cL = CovidInfos.fromJson(jsonDecode(s.data));
+                  var total = cL.cases;
+                  debugPrint('total: $total\n');
+                  return Column(
+                    children: <Widget>[
+                      CoronedGlobalCard(
+                        children: <Widget>[
+                          Text(
+                            'Statistiques Mondiales',
+                            style: Theme.of(context).textTheme.headline4,
+                          ),
+                          SizedBox(
+                            height: 8.0,
+                          ),
+                          Text(
+                            'CONTAMINÉS : ${cL.cases}',
+                            style: Theme.of(context)
+                                .textTheme
+                                .bodyText2
+                                .apply(color: Colors.black),
+                          ),
+                          Container(
+                            height: 8.0,
+                            width: MediaQuery.of(context).size.width / 2,
+                            decoration: BoxDecoration(
+                              color: aG.AppTheme.confirmedColorFill,
+                              border: Border.all(
+                                  color: aG.AppTheme.confirmedColorBorder),
+                            ),
+                          ),
+                          SizedBox(
+                            height: 8.0,
+                          ),
+                          Text(
+                            'MORTS : ${cL.deaths}',
+                            style: Theme.of(context)
+                                .textTheme
+                                .bodyText2
+                                .apply(color: Colors.black),
+                          ),
+                          Container(
+                            height: 8.0,
+                            width: MediaQuery.of(context).size.width / 2,
+                            decoration: BoxDecoration(
+                              color: aG.AppTheme.deathsColorFill,
+                              border: Border.all(
+                                  color: aG.AppTheme.deathsColorBorder),
+                            ),
+                          ),
+                          SizedBox(
+                            height: 8.0,
+                          ),
+                          Text(
+                            'GUÉRIS : ${cL.recovered}',
+                            style: Theme.of(context)
+                                .textTheme
+                                .bodyText2
+                                .apply(color: Colors.black),
+                          ),
+                          Container(
+                            height: 8.0,
+                            width: MediaQuery.of(context).size.width / 2,
+                            decoration: BoxDecoration(
+                              color: aG.AppTheme.recoveredColorFill,
+                              border: Border.all(
+                                  color: aG.AppTheme.recoveredColorBorder),
+                            ),
+                          ),
+                          SizedBox(
+                            height: 8.0,
+                          ),
+                        ],
+                      ),
+                    ],
+                  );
+                }
+                if (s.hasError) {
+                  return FailureIcon(fail: s.error);
+                }
+                if (s.connectionState != ConnectionState.done) {
+                  return Center(
+                    child: CircularProgressIndicator(),
+                  );
+                }
+                if (!s.hasData && s.connectionState == ConnectionState.done) {
+                  return FailureIcon(fail: 'No Data');
+                } else {
+                  return Center(
+                    child: CircularProgressIndicator(),
+                  );
+                }
+              },
             ),
           ),
         ],
@@ -91,51 +165,10 @@ class _DashBoardState extends State<DashBoard> {
     );
   }
 
-  List<Widget> _buildChildren() {
-    var children = <Widget>[
-      SizedBox(
-        height: 24.0,
-      ),
-    ];
-    children.add(
-      Center(
-        child: DropdownButton(
-          key: Key('Country List'),
-          onChanged: (String countryCode) {
-            setState(() {
-              _dropDownValue = countryCode;
-            });
-            call(countryCode);
-          },
-          hint: Text('Choisissez un pays'),
-          elevation: 2,
-          isExpanded: false,
-          value: _dropDownValue,
-          items: ['FR', 'US', 'UK', 'ERR']
-              .map((e) => DropdownMenuItem(
-                    child: Text('$e'),
-                    value: e,
-                  ))
-              .toList(),
-        ),
-      ),
-    );
-    children.add(
-      SizedBox(
-        height: 24.0,
-      ),
-    );
-    children.add(Text(
-      _apiResponse ?? ' ',
-      style: TextStyle(color: Colors.black),
-    ));
-    return children;
-  }
-
-  void call(String countryCode) {
+  void callApi() {
     locator
         .get<AppUtils>()
-        .getDataFromCountry(countryCode)
+        .getWorldLatestSituation()
         .then((value) => _apiResponseController.add(value))
         .catchError((e) => _apiResponseController.addError(e));
   }
