+++ lib/ui/widgets/widgets.dart
@@ -1,4 +1,4 @@
 export 'navbar_item/navbar_item.dart';
 export 'navigation_drawer/navigation_drawer.dart';
-export 'coroned_card.dart';
+export 'coroned_global_card.dart';
 export 'failure_icon.dart';
