+++ test/widget_test.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:cesi_covid_19_tracker/ui/pages/dashboard.dart';
+import 'package:cesi_covid_19_tracker/ui/pages/country_view.dart';
 
 void main() {
   final dropdownWidgetFinder = find.byKey(Key('Country List'));
   testWidgets('DashBoard dummy test', (WidgetTester tester) async {
-    final testWidget = buildTestableWidget(DashBoard());
+    final testWidget = buildTestableWidget(CountryView());
     // Build our app and trigger a frame.
     await tester.pumpWidget(testWidget);
     expect(dropdownWidgetFinder, findsOneWidget);
